+++ R/mod_Home.R
@@ -78,6 +78,7 @@ mod_Home_server <- function(id, i18n, Load_Slick_File, Slick_Object, Report){
     case_studies <- get_casestudies()
 
     output$load <- renderUI({
+      options(shiny.maxRequestSize = getOption("shiny.maxRequestSize", 200*1024^2))
       i18n <- i18n()
       tagList(
         shinydashboardPlus::box(width=3,