+++ NEWS.md
@@ -1,4 +1,13 @@
 
+# Slick 1.0.1
+
+- Added a `Misc` slot to all plot objects. 
+- Added option to show `mean` or `median` in the time series plots shown in the 
+  Slick app.
+- Fix issue with subsetting plots by Performance Indicators
+- Modify `plotTimeseries` so it can handle plots with no historical values
+- Fix to `plotQuilt(kable=TRUE)` for identical PI values across MPs
+
 # Slick 1.0.0
 
 `Slick` has been released on [CRAN](https://CRAN.R-project.org/package=Slick)